Add files via upload
In this task, i will be working with the Pandas and Matplotlib libraries to visualize data from a DataFrame. The goal is to create both a bar chart and a line chart to represent different aspects of the data. You will need to customize each chart by adding relevant labels for the axes, a descriptive title, and a legend to differentiate the data series. This exercise will help you understand how to create and customize visualizations in Python to effectively communicate data insights."